Update objects.js
Though the grey areas are automatically covered by last update there is still an anoying white area follow the sprite only in app mode and following the key event. I found it is the canvas of the sprite image. As you press the key it flashes there sometimes(not always). Make the sprite scale to 0.1 to make the canvas small and not easy to be feel. It can't be set to smaller size or the  line drawn with the move block would be biased.  Anyway very few people use keyboard to drive the sprite in  app mode(using the mouse doesn't trigger this problem). Maybe we could make it better in future versions.